+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/model/BSMetaProperty.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,6 +22,7 @@ import com.intellij.idea.plugin.hybris.system.bean.model.Property
 
 interface BSMetaProperty : BSMetaClassifier<Property>, BSTypedClassifier {
     val type: String?
+    val referencedType: String?
     val description: String?
     val isEquals: Boolean
     val isDeprecated: Boolean